+++ lib/whiteboard_data.dart
@@ -8,12 +8,12 @@ class WhiteboardData implements Serializable {
 
   @override
   void loadFromBytes(BinaryReader reader) {
-    var layerCount = reader.readUInt16();
+    var layerCount = reader.readUInt8();
     for (var i = 0; i < layerCount; i++) {
       layers.add(DrawingData()..loadFromBytes(reader));
     }
 
-    var textCount = reader.readUInt16();
+    var textCount = reader.readUInt8();
     for (var i = 0; i < textCount; i++) {
       texts.add(TextData()..loadFromBytes(reader));
     }
@@ -21,11 +21,11 @@ class WhiteboardData implements Serializable {
 
   @override
   void writeToBytes(BinaryWriter writer) {
-    writer.writeUInt16(layers.length);
+    writer.writeUInt8(layers.length);
     for (var layer in layers) {
       layer.writeToBytes(writer);
     }
-    writer.writeUInt16(texts.length);
+    writer.writeUInt8(texts.length);
     for (var text in texts) {
       text.writeToBytes(writer);
     }
